Increasing number of batches when single precision isn't large enough to store the batch size * N (for adjacency graph & CSR row in array.
Will eventually expose an argument through Python to allow the user to control the precision of the outputs (and thus the precision of the internal structures.). It's important to allow the user to set this explicitly because there's a huge tradeoff to be made between a larger number of much faster batches and a smaller number of much slower batches.