Check for any array interface in low-level types
If the object lacks `__array_interface__` and
`__cuda_array_interface__`, then it is low-level as we will need
metadata alongside the data to make sense of it. Though having these
doesn't mean it is not low-level as `Buffer` and `DeviceBuffer` support
`__cuda_array_interface__`. So we check these exceptions afterwards.