Fix SVR unit test parameter(#3294)
Linear SVR has the coef_ attribute in the python layer. In the C++ unit test the same vector is denoted by _w_, and it is defined as a linear combination of the support vectors

![image](https://user-images.githubusercontent.com/3671106/101908077-ce3d9e80-3bbb-11eb-98ff-e7be90828dde.png)

The number of elements in _w_ is n_cols. One of the SVR tests only defined 1 expected value for _w_, instead of the expected n_cols=2 values, which lead to accessing an uninitialized value. This would fail the test unless the value is accidentally zero initialized. Surprisingly this happened extremely rarely.

This PR fixes the expected value _w_exp_.

Authors:
  - Tamas Bela Feher <tfeher@nvidia.com>

Approvers:
  - Dante Gama Dessavre

URL: https://github.com/rapidsai/cuml/pull/3294