Fix brute force KNN distance metric issue (#3755)
The brute force KNN default distance metric switched from L2SqrtExpanded (or equivalent with FAISS metrics) to L2Unexpanded recently. I believe that it was a voluntary move to speedup calculations. However some calls to the raft::spatial::knn::brute_force_knn function in the cuML codebase required specifying a non-default metric. This is the case of distances that are not kept internally by an estimator but are instead provided to the end-user as results. This is also the case of features that specify a specific distance metric for their calculations. Please tell if you notice some other places where a non-default distance metric should have been specified.

Authors:
  - Victor Lafargue (https://github.com/viclafargue)

Approvers:
  - John Zedlewski (https://github.com/JohnZed)
  - Micka (https://github.com/lowener)
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuml/pull/3755