correct compliation of cuml c library (#3908)
Resolves two issues found when trying to build `cuml` locally. 

The first was that the `CUML_C_TARGET` had `FAISS` in the public link requirements, which doesn't look needed. This caused CMake configuration errors when trying to generate the `cuml-config.cmake` file. Solution was to move `FAISS` to be a private dependency ( brought in via `CUML_CPP_TARGET` )

The second was that the `CUML_C_TARGET`  tests need to make sure the `C` library is enabled. CUML doesn't enable this language by default as the `CUML_C_TARGET` builds using the `C++` compiler and exposes a `C` API via `extern C` blocks. 
I didn't go with a blanket enable `C` language as each language enabled, as that would increase the default requirements to build CUML when testing is disabled.

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/3908