Fixing UMAP reproducibility pytest failures in 11.4 by using random init for now (#4152)
This is side-stepping the spectral clustering bug for now. The spectral clustering using to be reproducible (and I'm fairly confident that it used to be returning the proper eigenpairs as well) but through various updates to the CTK, it seems to have gotten broken somewhere, potentially even in multiple places.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/4152